requirements.txt
discord.py == 0.16.12
youtube_dl
discord.ext
